Always allow to select fields to display
Make `fields` and `exclude_fields` consistently available as `ordering`.

This can improve performance, e.g. omit additional SQL queries to fetch
data for foreign keys.
