+++ pdc/apps/module/views.py
@@ -23,16 +23,6 @@ class UnreleasedVariantViewSet(viewsets.PDCModelViewSet):
     -d _data_ (a json string). or GUI plugins for
     browsers, such as ``RESTClient``, ``RESTConsole``.
 
-    ## Customize Output ##
-
-    There are two query params that you can use to customize your output.
-
-    `fields`:          string, can be set multiple times, to demand what fields you want to include;
-
-    `exclude_fields`:  string, can be set multiple times, to demand what fields you do NOT want.
-
-    __NOTE__: If both given, `exclude_fields` *rules* `fields`.
-
     """
     model = UnreleasedVariant
     queryset = UnreleasedVariant.objects.all().order_by('variant_uid')
